refactor search logic
